refactor : 하트아이콘 사이즈 변경 기능 추가 #34
