+++ app/(home)/(archive)/(for-you)/page.tsx
@@ -1,11 +1,11 @@
-import HeartButton from '../../../../components/buttons/HeartButton';
+import HeartButton from '../../../../components/animations/HeartButton';
 
 export default function ForYouPage() {
 
   return (
     <section className="ForYouContainer w-full h-full">
       ForYouContainer 입니다.
-      <HeartButton/>
+      <HeartButton width='21' height='19'/>
     </section>
   );
 }
