+++ components/animations/PlusButton.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import { motion, AnimatePresence } from "framer-motion";
+import { motion } from "framer-motion";
 
 export default function PlusButton() {
   const [isClick, setIsClick] = useState(false);
