+++ constants/interface.ts
@@ -6,3 +6,8 @@ export interface CategoryElement {
 export interface CategoryBarProps {
   elements: CategoryElement[];
 }
+
+export interface SizeProps {
+  width: string;
+  height: string;
+}
